requirements.txt
flask
deepface
opencv-python
pillow
numpy
tqdm
werkzeug
